+++ frontend/src/metabase/search/components/UserNameDisplay/index.ts
@@ -1,0 +1,1 @@
+export * from "./UserNameDisplay";
