--- frontend/src/metabase/search/components/filters/CreatedByFilter/CreatedByContent/index.ts
@@ -1,1 +1,0 @@
-export * from "./CreatedByContent";
