--- frontend/src/metabase/search/components/filters/CreatedByFilter/CreatedByDisplay/index.ts
@@ -1,1 +1,0 @@
-export * from "./CreatedByDisplay";
